// frontend/src/api/generated/model/transportSpecType.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */

export type TransportSpecType = typeof TransportSpecType[keyof typeof TransportSpecType];


export const TransportSpecType = {
  native: 'native',
  'sing-box': 'sing-box',
  'sing-box-vless-reality': 'sing-box-vless-reality',
} as const;